Combine the tokenizer all together into one minifier
