+++ Sources/Utilities/Minifier/Extensions/Minifier+Character.swift
