+++ Sources/Utilities/Minifier/Tokenization/Stylesheet/Stylesheet.swift
@@ -1,4 +1,4 @@
-internal class Tokenizer {
+internal class Stylesheet {
     
     /// A enumeration of different states of the minifier
     ///
@@ -145,14 +145,14 @@ internal class Tokenizer {
         
         if character.isNewline {
             
-            self.emit(token: WhitespaceToken(type: .return ,value: String(character)))
+            self.emit(token: Stylesheet.WhitespaceToken(type: .return ,value: String(character)))
             
             return .code
         }
         
         if character.isWhitespace {
             
-            self.emit(token: WhitespaceToken(type: .whitespace, value: String(character)))
+            self.emit(token: Stylesheet.WhitespaceToken(type: .whitespace, value: String(character)))
             
             return .code
         }
@@ -164,35 +164,35 @@ internal class Tokenizer {
         
         if character.isPeriod {
             
-            self.assign(token: SelectorToken(type: .class, value: ""))
+            self.assign(token: Stylesheet.SelectorToken(type: .class, value: ""))
             
             return .selector
         }
         
         if character.isColon {
             
-            self.assign(token: SelectorToken(type: .root, value: ""))
+            self.assign(token: Stylesheet.SelectorToken(type: .root, value: ""))
             
             return .selector
         }
         
         if character.isNumberSign {
             
-            self.assign(token: SelectorToken(type: .id, value: ""))
+            self.assign(token: Stylesheet.SelectorToken(type: .id, value: ""))
             
             return .selector
         }
         
         if character.isLetter {
             
-            self.assign(token: SelectorToken(type: .element, value: ""))
+            self.assign(token: Stylesheet.SelectorToken(type: .element, value: ""))
             
             return .selector
         }
         
         if character.isBracket {
             
-            self.emit(token: FormatToken(type: .punctuator, value: String(character)))
+            self.emit(token: Stylesheet.FormatToken(type: .punctuator, value: String(character)))
             
             return .declaration
         }
@@ -207,7 +207,7 @@ internal class Tokenizer {
         
         if character.isSolidus {
             
-            self.assign(token: CommentToken(type: .line, value: ""))
+            self.assign(token: Stylesheet.CommentToken(type: .line, value: ""))
         
             // ignore character
             return .comment
@@ -215,7 +215,7 @@ internal class Tokenizer {
         
         if character.isAsterisk {
             
-            self.assign(token: CommentToken(type: .block, value: ""))
+            self.assign(token: Stylesheet.CommentToken(type: .block, value: ""))
         
             // ignore character
             return .comment
@@ -299,14 +299,14 @@ internal class Tokenizer {
         
         if character.isBracket {
             
-            self.emit(token: FormatToken(type: .punctuator, value: String(character)))
+            self.emit(token: Stylesheet.FormatToken(type: .punctuator, value: String(character)))
             
             return .code
         }
         
         if character.isLetter {
         
-            self.assign(token: PropertyToken(type: .standard, value: String(character)))
+            self.assign(token: Stylesheet.PropertyToken(type: .standard, value: String(character)))
             
             return .property
         }
@@ -332,7 +332,7 @@ internal class Tokenizer {
             
             self.emit()
             
-            self.emit(token: FormatToken(type: .punctuator, value: String(character)))
+            self.emit(token: Stylesheet.FormatToken(type: .punctuator, value: String(character)))
             
             return .value
         }
@@ -362,7 +362,7 @@ internal class Tokenizer {
             
             self.emit()
             
-            self.emit(token: FormatToken(type: .terminator, value: String(character)))
+            self.emit(token: Stylesheet.FormatToken(type: .terminator, value: String(character)))
             
             return .declaration
         }
@@ -371,7 +371,7 @@ internal class Tokenizer {
             token.value.append(character)
             
         } else {
-            self.assign(token: ValueToken(type: .keyword, value: String(character)))
+            self.assign(token: Stylesheet.ValueToken(type: .keyword, value: String(character)))
         }
         
         return .value
@@ -387,5 +387,243 @@ internal class Tokenizer {
         self.log(function: #function, character: character)
         
         return .mediaquery
+    }
+}
+
+extension Stylesheet {
+
+    internal class WhitespaceToken: Token {
+        
+        /// A enumeration of the variations of comment tokens
+        internal enum TokenType {
+            
+            /// Indicates a single line comment
+            case whitespace
+            
+            /// Indicates a multiline comment
+            case `return`
+        }
+        
+        /// The type of the token
+        internal var type: TokenType
+        
+        /// The value of the token
+        internal var value: String
+        
+        /// Initiates a comment token
+        internal init(type: TokenType, value: String) {
+            
+            self.type = type
+            self.value = value
+        }
+        
+        internal func present() -> String {
+            
+            switch type {
+            case .return:
+                return "\r\n"
+                
+            case .whitespace:
+                return " "
+            }
+        }
+    }
+
+    /// A type that represents a css comment
+    internal class CommentToken: Token {
+        
+        /// A enumeration of the variations of comment tokens
+        internal enum TokenType {
+            
+            /// Indicates a single line comment
+            case line
+            
+            /// Indicates a multiline comment
+            case block
+        }
+        
+        /// The type of the token
+        internal var type: TokenType
+        
+        /// The value of the token
+        internal var value: String
+        
+        /// Initiates a comment token
+        internal init(type: TokenType, value: String) {
+            
+            self.type = type
+            self.value = value
+        }
+        
+        /// Minifies a comment token
+        internal func present() -> String {
+            
+            switch type {
+            case .line:
+                return "//\(value)"
+                
+            case .block:
+                return "/*\(value)*/"
+            }
+        }
+    }
+
+    /// A type that represents a css comment
+    internal class SelectorToken: Token {
+        
+        /// A enumeration of the variations of comment tokens
+        internal enum TokenType {
+            
+            /// Indicates the set is about an element
+            case element
+            
+            /// Indicates the set is about a class
+            case `class`
+            
+            /// Indicates the set is about a id
+            case id
+            
+            /// Indicates the set is about a root
+            case root
+        }
+        
+        /// The type of the token
+        internal var type: TokenType
+        
+        /// The value of the token
+        internal var value: String
+        
+        /// Initiates a comment token
+        internal init(type: TokenType, value: String) {
+            
+            self.type = type
+            self.value = value
+        }
+        
+        /// Minifies a comment token
+        internal func present() -> String {
+            
+            switch type {
+            case .element:
+                return "\(value)"
+                
+            case .class:
+                return ".\(value)"
+                
+            case .id:
+                return "#\(value)"
+                
+            case .root:
+                return ":\(value)"
+            }
+        }
+    }
+
+    /// A type that represents a format control token
+    internal class FormatToken: Token {
+        
+        /// A enumeration of the variation of format tokens
+        internal enum TokenType {
+            
+            /// Indicates a punctiation
+            case punctuator
+            
+            /// Indicates a line terminator character
+            case terminator
+            
+            /// Indicates a line operator character
+            case `operator`
+        }
+        
+        /// The type of the token
+        internal var type: TokenType
+        
+        /// The value of the token
+        internal var value: String
+        
+        /// Initiates a format token
+        internal init(type: TokenType, value: String) {
+            
+            self.type = type
+            self.value = value
+        }
+        
+        /// Minifies a format token
+        internal func present() -> String {
+            return value
+        }
+    }
+
+    /// A type that represents a format control token
+    internal class PropertyToken: Token {
+        
+        /// A enumeration of the variation of format tokens
+        internal enum TokenType {
+            
+            /// Indicates a punctiation
+            case standard
+            
+            /// Indicates a line terminator character
+            case custom
+        }
+        
+        /// The type of the token
+        internal var type: TokenType
+        
+        /// The value of the token
+        internal var value: String
+        
+        /// Initiates a format token
+        internal init(type: TokenType, value: String) {
+            
+            self.type = type
+            self.value = value
+        }
+        
+        /// Minifies a format token
+        internal func present() -> String {
+            
+            switch type {
+            case .standard:
+                return "\(value)"
+                
+            case .custom:
+                return "--\(value)"
+            }
+        }
+    }
+
+    /// A type that represents a format control token
+    internal class ValueToken: Token {
+        
+        /// A enumeration of the variation of format tokens
+        internal enum TokenType {
+            
+            /// Indicates a punctiation
+            case keyword
+            
+            /// Indicates a line terminator character
+            case numeric
+            
+            case function
+        }
+        
+        /// The type of the token
+        internal var type: TokenType
+        
+        /// The value of the token
+        internal var value: String
+        
+        /// Initiates a format token
+        internal init(type: TokenType, value: String) {
+            
+            self.type = type
+            self.value = value
+        }
+        
+        /// Minifies a format token
+        internal func present() -> String {
+            return value
+        }
     }
 }
