+++ Sources/Utilities/Minifier/Tokenization/Token.swift
@@ -1,0 +1,7 @@
+/// A Type that represents a javascript token
+internal protocol Token {
+    
+    var value: String { get set }
+    
+    func present() -> String
+}
