+++ Tests/UtilitiesTests/MinifierTests.swift
@@ -1,5 +1,5 @@
 import XCTest
-import CSSMinifier
+import Minifier
 
 final class UtilitiesTests: XCTestCase {
     
@@ -9,6 +9,15 @@ final class UtilitiesTests: XCTestCase {
                     .test {}
                     """
         
-        XCTAssertEqual(Minifier().minify(css), ".test{}")
+        XCTAssertEqual(Minifier().minify(css: css), ".test{}")
+    }
+    
+    func testDeclaration() throws {
+        
+        let css =   """
+                    var test = "test";
+                    """
+        
+        XCTAssertEqual(Minifier().minify(js: css), "var test='test';")
     }
 }
